Fixed and folder named problem for Github and created shared folder (1) Try to fix Github has two "frontEnd" folders (2) Moved shared models and services into "shared" folder (3) Unified naming method
