+++ FrontEnd1/src/app/home/home.component.scss
