+++ FrontEnd1/src/app/rug-detail/rug-detail.component.scss
