+++ FrontEnd1/src/app/rug-list/rug-list.component.scss
