+++ FrontEnd1/src/app/shared/services/rug.service.ts
@@ -1,7 +1,7 @@
-import { Injectable } from '@angular/core';
 import { Http, Response, Headers, RequestOptions } from '@angular/http';
+import { Injectable } from '@angular/core';
 import 'rxjs/add/operator/map';
-import { Rug } from './rug';
+import { Rug } from '../models/rug.model';
 
 @Injectable()
 export class RugService {
