+++ FrontEnd1/src/tsconfig.app.json
